put interface noder into its own file
